# lib/api-zod/src/generated/types/reasoningMetric.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReasoningMetric {
  label: string;
  value: string;
  /** @nullable */
  detail?: string | null;
}